+++ Win/WinPin.cpp
@@ -269,7 +269,7 @@ void WinPin::mouseRelease(QMouseEvent* event)
             showToolMain();
         }
         auto span = event->globalPosition() - posPress;
-        for (size_t i = 0; i < shapes.size(); i++)
+        for (int i = 0; i < shapes.size(); i++)
         {
             auto shapeText = qobject_cast<ShapeText*>(shapes[i]);
             if (shapeText) {
@@ -305,10 +305,10 @@ void WinPin::onMouseWheel(const int& delta)
     if (state == State::start) {
         float scale;
         if (delta > 0) {
-            scale = 1.1;
+            scale = 1.1f;
         }
         else {
-            scale = 0.9;
+            scale = 0.9f;
         }
         //imgRect.setXYWH(padding, padding, imgRect.width() * scale, imgRect.height() * scale);
         //auto tm = ToolMain::Get();
